API: add TerraFusion.Consciousness project reference to stabilize DI and integration tests; backend builds and test pipeline pass locally

+++ backend/TerraFusion.AI/Services/AIAssistantService.cs
@@ -235,7 +235,7 @@ namespace TerraFusion.AI.Services
             {
                 MessageId = Guid.NewGuid().ToString(),
                 Content = content,
-                Confidence = swarmResponse.ConfidenceScore,
+                Confidence = (double)swarmResponse.ConfidenceScore,
                 Suggestions = suggestions,
                 Metadata = swarmResponse.Results,
                 Timestamp = DateTime.UtcNow,
